requirements.txt
asgiref==3.5.2
autopep8==2.0.1
awscli==1.27.19
boto3==1.26.19
botocore==1.29.19
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
cryptography==38.0.4
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.4
django-allauth==0.52.0
django-cors-headers==3.13.0
django-rest-knox==4.2.0
django-user-agents==0.4.0
djangorestframework==3.14.0
docutils==0.16
future==0.18.3
gdrive==0.1.5
geocoder==1.38.1
google-api-core==2.11.0
google-api-python-client==2.43.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.58.0
greenlet==2.0.2
gspread==5.8.0
httplib2==0.21.0
idna==3.4
jeepney==0.8.0
jmespath==1.0.1
JPype1==1.4.0
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
pandas==1.5.2
pdfkit==1.0.0
Pillow==9.3.0
protobuf==4.21.12
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pygdrive3==0.8.1
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
PyYAML==5.4.1
ratelim==0.1.6
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.6.0
SecretStorage==3.3.3
six==1.16.0
SQLAlchemy==2.0.1
sqlparse==0.4.3
stripe==5.2.0
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.14
user-agents==2.2.0
versioneer==0.28
